Represent tiles by objects instead of numbers
Previously, tetrominoes were represented by zeros and ones. However,
before merging them into the playfield, the ones needed to be converted
into a unique number for each tetromino type, so the tiles would be
colored differently. This meant that the Tetris core needed to have
some kind of color map to color the tetrominoes properly, and then the
inverse of that map had to be used in Tetris canvas, to convert the
numbers in the playfield into colors.

The playfield now stores zeros and objects. Each tetromino
has a unique object reference in its `value` property, and coloring the
tiles is done inside Tetris canvas by using an object to string map.